- fix event policy

+++ app/Http/Controllers/EventController.php
@@ -34,8 +34,8 @@ use App\Models\UserShiftCalendarFilter;
 use App\Support\Services\CollisionService;
 use App\Support\Services\NewHistoryService;
 use App\Support\Services\NotificationService;
-use Barryvdh\Debugbar\Facades\Debugbar;
 use Carbon\Carbon;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
@@ -318,6 +318,8 @@ class EventController extends Controller
     //phpcs:ignore Generic.Metrics.CyclomaticComplexity.TooHigh
     public function storeEvent(EventStoreRequest $request): CalendarEventResource
     {
+        $this->authorize('create', Event::class);
+
         $firstEvent = Event::create($request->data());
         $this->adjoiningRoomsCheck($request, $firstEvent);
         if ($request->get('projectName')) {
@@ -506,7 +508,7 @@ class EventController extends Controller
                 }
             }
         }
-        $this->authorize('create', Event::class);
+
         $collisionsCount = $this->collisionService->getCollision($request, $event)->count();
         if ($collisionsCount > 0) {
             $collisions = $this->collisionService->getConflictEvents($request);
@@ -730,10 +732,15 @@ class EventController extends Controller
         }
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     //@todo: fix phpcs error - refactor function because complexity is rising
     //phpcs:ignore Generic.Metrics.CyclomaticComplexity.TooHigh
     public function updateEvent(EventUpdateRequest $request, Event $event): CalendarEventResource
     {
+        $this->authorize('update', $event);
+
         if (!$request->noNotifications) {
             $projectManagers = [];
             $this->notificationService->setNotificationKey($this->notificationKey);
@@ -871,8 +878,6 @@ class EventController extends Controller
             $this->notificationService->createNotification();
         }
 
-        $this->authorize('update', $event);
-
         $oldEventDescription = $event->description;
         $oldEventRoom = $event->room_id;
         $oldEventProject = $event->project_id;
@@ -988,8 +993,13 @@ class EventController extends Controller
         }
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function answerOnEvent(Event $event, Request $request): void
     {
+        $this->authorize('update', $event);
+
         $event->comments()->create([
             'user_id' => Auth::id(),
             'comment' => $request->comment,
@@ -1073,8 +1083,13 @@ class EventController extends Controller
         }
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function acceptEvent(Request $request, Event $event): RedirectResponse
     {
+        $this->authorize('update', $event);
+
         $event->occupancy_option = false;
         $notificationTitle = 'Raumanfrage bestätigt';
         $this->history->createHistory($event->id, 'Raum bestätigt');
@@ -1141,10 +1156,15 @@ class EventController extends Controller
         return Redirect::back();
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     //@todo: fix phpcs error - refactor function because complexity is rising
     //phpcs:ignore Generic.Metrics.CyclomaticComplexity.TooHigh
     public function declineEvent(Request $request, Event $event): void
     {
+        $this->authorize('update', $event);
+
         $projectManagers = [];
         $roomId = $event->room_id;
         $room = $event->room()->first();
@@ -1380,9 +1400,12 @@ class EventController extends Controller
         ]);
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function destroyShifts(Event $event): RedirectResponse
     {
-        Debugbar::info("Deleting shifts of event $event->id");
+        $this->authorize('update', $event);
 
         $event->shifts()->delete();
         $event->timeline()->delete();
@@ -1390,9 +1413,13 @@ class EventController extends Controller
         return redirect()->back();
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function destroy(Event $event): RedirectResponse
     {
         $this->authorize('delete', $event);
+
         if (!empty($event->project_id)) {
             $eventProject = $event->project()->first();
             $projectHistory = new NewHistoryService(Project::class);
@@ -1463,6 +1490,9 @@ class EventController extends Controller
         return Redirect::back();
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function destroyByNotification(Event $event, Request $request): void
     {
         $this->authorize('delete', $event);
@@ -1543,11 +1573,17 @@ class EventController extends Controller
         }
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function forceDelete(int $id): RedirectResponse
     {
         $event = Event::onlyTrashed()->findOrFail($id);
+
         $this->authorize('delete', $event);
+
         $event->forceDelete();
+
         broadcast(new OccupancyUpdated())->toOthers();
 
         return Redirect::route('events.trashed')->with('success', 'Event deleted');
